Split shortcut session effects
